Pass the CPU index as a parameter to startup.
This addition is in support of future multicore support in
coreboot. It also will allow us to remove some asssembly code.

The CPU "index" -- i.e., its order in the sequence in which
cores are brought up, NOT its APIC id -- is passed into the
secondary start. We modify the function to specify regparm(0).
We also take this opportunity to do some cleanup:
indexes become unsigned ints, not unsigned longs, for example.

Build and boot on a multicore system, with pcserial enabled.

Capture the output. Observe that the messages
Initializing CPU #0
Initializing CPU #1
Initializing CPU #2
Initializing CPU #3
appear exactly as they do prior to this change.

+++ src/arch/x86/include/arch/cpu.h
@@ -160,7 +160,7 @@ struct cpu_driver *find_cpu_driver(struct device *cpu);
 
 struct cpu_info {
 	device_t cpu;
-	unsigned long index;
+	unsigned int index;
 };
 
 static inline struct cpu_info *cpu_info(void)
